Consolidate finance feature types
