+++ web/src/features/finance/RateSnapshotsWorkspace.tsx
@@ -35,13 +35,14 @@ import {
   SnapshotNavigator,
   SnapshotSelectorToolbar,
 } from "./SnapshotChrome";
-import type { RateRowState, RateSnapshotFormMode } from "./types";
 import {
   isoToDateTimeLocal,
   localDateTimeToIso,
   nowDateTimeLocal,
   rateEntryEquation,
   rateSnapshotLabel,
+  type RateRowState,
+  type RateSnapshotFormMode,
 } from "./utils";
 import { useFinanceAssetSource } from "./useFinanceAssetSource";
